chore: rename types to hash_trait and add pedersen file
