+++ crates/committer/src/hash.rs
@@ -1,1 +1,2 @@
-pub mod types;
+pub mod hash_trait;
+pub mod pedersen;
