+++ crates/committer/src/patricia_merkle_tree/filled_node.rs
@@ -1,5 +1,5 @@
 use crate::patricia_merkle_tree::types::{EdgeData, LeafDataTrait};
-use crate::{hash::types::HashOutput, types::Felt};
+use crate::{hash::hash_trait::HashOutput, types::Felt};
 // TODO(Nimrod, 1/6/2024): Swap to starknet-types-core types once implemented.
 #[allow(dead_code)]
 pub(crate) struct ClassHash(pub Felt);
