+++ crates/committer/src/patricia_merkle_tree/types_test.rs
@@ -1,4 +1,5 @@
-use crate::hash::types::{HashFunction, HashInputPair, HashOutput, PedersenHashFunction};
+use crate::hash::hash_trait::{HashFunction, HashInputPair, HashOutput};
+use crate::hash::pedersen::PedersenHashFunction;
 use crate::patricia_merkle_tree::filled_node::{BinaryData, NodeData};
 use crate::patricia_merkle_tree::types::TreeHashFunction;
 use crate::patricia_merkle_tree::types::{
